% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign-detect-time.R
\name{assign_detection_time}
\alias{assign_detection_time}
\title{Takes a simulated infection time and maps it to an actual detection time from the data}
\usage{
assign_detection_time(sim_infection_time, raw_data_times)
}
\arguments{
\item{sim_infection_time}{Infection time from a simulation}

\item{raw_data_times}{Sampling times from the data where detection would be detect}
}
\value{
A time that matches the detection/observation times in the data
}
\description{
Takes a simulated infection time and maps it to an actual detection time from the data
}
\examples{
assign_detection_time(7.2, c(0,13,21,33))
assign_detection_time(7.2, c(0,13,21,33))
}

