% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_likelihood.R
\name{calc_censor_likelihood}
\alias{calc_censor_likelihood}
\title{Calculates the interval censored likelihood when there is censoring related to the infection time}
\usage{
calc_censor_likelihood(exposure_series, exposure_times, censor_start,
  censor_end, hazard_risk, clr, loglik = T)
}
\arguments{
\item{exposure_series}{The vector of repeated pathogen exposure values, should be pre-processed}

\item{exposure_times}{The times of exposures}

\item{censor_start}{Start of the censor period (exposure up to this time are survived)}

\item{censor_end}{End of the censor period (infection occured in exposures between censor_start and censor_end)}

\item{hazard_risk}{The per time risk of exposure during persistence}

\item{clr}{The clearance rate of exposure}

\item{loglik}{Whether to return log (default TRUE)}
}
\value{
Returns a survival likelihood for interval censored outcome (see \code{\link{calc_risk_mult}} for calculations)
}
\description{
For interval censoring with left censor time Tl and right censor time Tr, this returns S(Tl) - S(Tr).
}

