% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_exposure.R
\name{calc_current_exposure}
\alias{calc_current_exposure}
\title{Calculates the remaining exposure at given time}
\usage{
calc_current_exposure(exposure_series, exposure_times, final_time, clr,
  na.zero = F)
}
\arguments{
\item{exposure_series}{The values of repeated pathogen exposures}

\item{exposure_times}{The times of exposures}

\item{final_time}{The time to calculate remaining exposure at}

\item{clr}{The clearance rate of exposure}
}
\value{
remaining exposure and end of time frame (given by final_times)
}
\description{
This function takes a set of exposures and exposures times (or a single exposure) and calculates the remaining level at the end of a certain time frame
The exposures decay exponentially (see \code{\link{calc_risk}}) and if end_time -> infty, the remaining exposure will be zero
}
\examples{
calc_current_exposure_cpp(rep(1e7, 5), 0:4, 5, 2)
}

