% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_current_exposure_cpp}
\alias{calc_current_exposure_cpp}
\title{This is a C++ version of the current remaining exposure function}
\usage{
calc_current_exposure_cpp(exposure, start_time, final_day, clr)
}
\description{
This is a C++ version of the current remaining exposure function
}

