% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{calc_risk}
\alias{calc_risk}
\title{Calculate risk from one pathogen exposure}
\usage{
calc_risk(exposure, survival_time = Inf, hazard_risk, clr)
}
\arguments{
\item{exposure}{The initial level of the pathogen exposure}

\item{survival_time}{The range of time to calculate the risk over (default infinity)}

\item{hazard_risk}{The per time risk of exposure during persistence}

\item{clr}{The clearance rate of exposure}
}
\value{
risk calculation over the time specified in survival time
}
\description{
This function takes a pathogen exposure and calculates the risk of infection over its persistence within host
}
\details{
The risk is derived using a surival function.  exposure = p, survival_time = T, hazard_risk = \eqn{\beta}, and clearance rate = c.
Note: every calculation assumes that time starts at 0

The pathogen is assumed to be cleared exponentially (\eqn{P(0) = p})
\deqn{\frac{\mathrm{d}P}{\mathrm{d}t} = -cP}{dP/dt = -cP}

The pathogens pose a risk of infection during their persistance that can be used to calculate the survival function
\deqn{\lambda(t) = \beta * P(t)}
\deqn{S(T) =  \mathrm{e}^{-\int_0^T{\lambda(t)\mathrm{d}t}}}{S(T) =  exp(-integral(0, T) \lambda(t)dt)}
\deqn{S(T) = \mathrm{e}^{-\frac{\beta}{c} * p * (1 - \mathrm{e}^{-c(T)})}}{S(t) = exp(-\beta/c * p * (1 - exp(-cT))}
which simplifies as \eqn{T \rightarrow \infty}{T -> Inf} to
\deqn{S(T) = \mathrm{e}^{-\frac{\beta}{c}p}}{S(T) = exp(-\beta/c * p)}

The risk is \eqn{1 - S(T)}.
}
\examples{
calc_risk(1e8, 1, 1e-8, 1)
calc_risk(exposure = 1e8, hazard_risk = 1e-8, clr = 1)
}
\seealso{
\code{\link{calc_risk_mult} for mulitple exposures.}
}

