% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{calc_risk_mult}
\alias{calc_risk_mult}
\title{Calculate risk from mulitple pathogen exposures}
\usage{
calc_risk_mult(exposure_series, exposure_times, survival_time = Inf,
  hazard_risk, clr, na.zero = F, log.surv = F)
}
\arguments{
\item{exposure_series}{The values of repeated pathogen exposures}

\item{exposure_times}{The times of exposures}

\item{survival_time}{the final time to calculate survival to (default infinity)}

\item{hazard_risk}{The per time risk of exposure during persistence}

\item{clr}{The clearance rate of exposure}
}
\value{
risk calculation over the time specified in survival time
}
\description{
This function takes mulitple pathogen exposure and calculates the risk of infection over their persistence within host
}
\details{
The risk for a single exposure is derived in \code{\link{calc_risk}}.  The accumulation of exposure assuming exponential clearance is independent.
\deqn{S(T) = \mathrm{e}^{-\frac{\beta}{c}\sum\limits_{i=1}^n{p_i(1 - \mathrm{e}^{-c(T - t_i)})}}}{S(t) = exp(-\beta/c *sum(i = 1..n) p_i *(1 - exp(-c(T - t_i))}
which simplifies as \eqn{T \rightarrow \infty}{T -> Inf} to
\deqn{S(T) = \mathrm{e}^{-\frac{\beta}{c}\sum\limits_{i=1}^n{p_i}}}{S(T) = exp(-\beta/c *sum(i = 1..n) p_i)}
}
\examples{
##Single exposure examples
calc_risk_mult(1e8, 0, 1, 1e-8, 1) # should return same as calc_risk(1e8, 1, 1e-8, 1)
calc_risk_mult(1e8, 1, 2, 1e-8, 1) # should return same as calc_risk(1e8, 1, 1e-8, 1)
## multiple exposures
exposures = c(1e7, 1e9, 1e5); exp_times = c(0, 1, 2)
calc_risk_mult(exposures, exp_times, 3, hazard_risk = 1e-9, clr = 1)
calc_risk_mult(exposures, exp_times, hazard_risk = 1e-9, clr = 1) #total risk
1-exp(-1e-9/1 * sum(exposures))
}
\seealso{
\code{\link{calc_risk} for single exposures.}
}

