% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_risk_mult_cpp}
\alias{calc_risk_mult_cpp}
\title{This is a C++ version of the risk calculation function}
\usage{
calc_risk_mult_cpp(exposure, start_time, final_day, hazard_risk, clr)
}
\description{
This is a C++ version of the risk calculation function
}

