% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_handling.R
\name{check_exposure_input}
\alias{check_exposure_input}
\title{Looks at exposure series and exposure times for common errors}
\usage{
check_exposure_input(exposure_series, exposure_times, na.zero = F)
}
\description{
Looks at exposure series and exposure times for common errors
}
\keyword{internal}

