% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process-survlik-data.R
\name{likelihood_setup_call}
\alias{likelihood_setup_call}
\title{Processes data for and sends data to survival likelihood function}
\usage{
likelihood_setup_call(exposure_data, betat, clr, lag, end_time,
  censor_start_time)
}
\arguments{
\item{exposure_data}{A data frame that contains time and exposures}

\item{clr}{The clearance rate of exposure}

\item{hazard_risk}{The per time risk of exposure during persistence}

\item{censor_start}{Start of the interval censored outcome, should be passed in from the data}
}
\value{
Returns a log survival likelihood (see \code{\link{calc_risk_mult}}) for uninfected and an interval censored likelihood
(see \code{\link{calc_censor_likelihood}}) for infected
}
\description{
Processes data for and sends data to survival likelihood function
}

