% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_likelihood.R
\name{log_likelihood_fun}
\alias{log_likelihood_fun}
\title{Calculates the likelihood of the exposure data}
\usage{
log_likelihood_fun(infected, exposure_series, exposure_times, end_time,
  hazard_risk, clr = 2, censor_start = NA)
}
\arguments{
\item{exposure_series}{The vector of repeated pathogen exposure values, should be pre-processed}

\item{exposure_times}{The times of exposures}

\item{end_time}{Final time: either censor time (infected = 0) or infection time (infected = 1).  If there is a lag, process it that first.}

\item{hazard_risk}{The per time risk of exposure during persistence}

\item{clr}{The clearance rate of exposure}

\item{censor_start}{Start of the interval censored data (end_time is the end)}
}
\value{
Returns a log survival likelihood (see \code{\link{calc_risk_mult}}) for uninfected and an interval censored likelihood
(see \code{\link{calc_censor_likelihood}}) for infected
}
\description{
Calculates the likelihood of the exposure data
}

