% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run-simulations.R
\name{run_sample_times_simulations}
\alias{run_sample_times_simulations}
\title{Runs multiple simulations for infection time by ID and given infection detection times}
\usage{
run_sample_times_simulations(nsims = 100, raw_exposure_data, raw_sample_times,
  exposure_rate, hazard_risk, clr = 2, lag = 0, return_KM = F,
  parallel = F)
}
\arguments{
\item{raw_exposure_data}{The data.frame time containing series of all possible exposures (must include a day and count name) for all subjects (FamilyID)}

\item{raw_sample_times}{The sampling days that infection would actually be observed for each subject}

\item{exposure_rate}{The rate of exposures to be sampled from raw_exposure_data}

\item{hazard_risk}{The per time hazard risk for the hazard function}

\item{clr}{The exponential pathogen clearance rate}

\item{lag}{The minimum amount of time between infection and when it could be detected (lag = 0 means infection is instantly detectable)}

\item{return_KM}{return the Kaplan-Meier estimates from all the runs (if false, just returns raw data)}

\item{parallel,}{Use parallel computation for simulations across nsim}

\item{nsim}{The amount of simulations to run}
}
\value{
Either a data.frame of infection time simulations for each family matched to sample times or a Kaplan-Meier object
}
\description{
This function takes data of multiple subjects with repeated exposure data and runs mulitple stochastic simulations estimating infection time
infection time that is linked to the given detection time. For more flexibility on returning exact infection times,
use TBD.
}

