% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample-exposures.R
\name{sample_exposure_data}
\alias{sample_exposure_data}
\title{Takes a series of exposures and times and samples from them based on an exposure rate}
\usage{
sample_exposure_data(raw_exposure_data, exposure_rate)
}
\arguments{
\item{raw_exposure_data}{The subsetted data.frame containing time (days) and exposure (count) for one subject}

\item{exposure_rate}{The rate of sampling from the raw_exposure_data}
}
\value{
A data.frame containing only the sampled exposure times
}
\description{
Effectively maps a continuum of exposure to a discrete set of actual exposures that could cause infection.
}

