% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate-infections.R
\name{simulate_infection_sample_times}
\alias{simulate_infection_sample_times}
\title{Simulates infection time for one exposure time series based on given infection detection times}
\usage{
simulate_infection_sample_times(exposure_ts, sample_times, hazard_risk, clr,
  lag = 0, debug = F)
}
\arguments{
\item{exposure_ts}{The data.frame time series of exposure (must include a day and count name)}

\item{sample_times}{The sampling days that infection would actually be observed}

\item{hazard_risk}{The per time hazard risk for the hazard function}

\item{clr}{The exponential pathogen clearance rate}

\item{lag}{The minimum amount of time between infection and when it could be detected (lag = 0 means infection is instantly detectable)}
}
\value{
An infection time that matches a time given in sample_times
}
\description{
This function takes a time series of exposures, exposure times, and infection detection times and returns a stochastic
infection time that is linked to the given detection time. For more flexibility on returning exact infection times,
use TBD.
}

