% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_infection_times_cpp}
\alias{simulate_infection_times_cpp}
\title{This is a C++ version of the simulation function, that will replace the while loop in the simulation call
 Assigning infection time to sample time is done later}
\usage{
simulate_infection_times_cpp(exposure_set, time_set, hazard_risk, clr)
}
\description{
This is a C++ version of the simulation function, that will replace the while loop in the simulation call
 Assigning infection time to sample time is done later
}

